Add clock module and change tachometry module to use input change notification instead of event capture

+++ Power_Board/src/motor.h
@@ -79,10 +79,10 @@ void motor_init(MotorChannel channel);
 /// Initialize motor feedback (tachometry)
 void motor_tach_init();
 
-/// Get the last-measured period of a motor in units of 16-microseconds
+/// Get the last-measured period of a motor in units of clock ticks
 /// @param channel which motor?
-/// @return Period of the motor in units of 16 microseconds. If the motor period is long, returns 0.
-float motor_tach_get_period(MotorChannel channel);
+/// @return Period of the motor in units of clock ticks.
+int64_t motor_tach_get_period(MotorChannel channel);
 
 /// Tell the motor controller what to do with the motor
 /// @param channel Which motor to update?
